Add files via upload
Big data dump of forwards and defense. EvaluatedPlayer is the player of question and 'Player' is the player related to the similarity score which includes all future seasons of this player to get an idea of how their career progressed. Named MILLER because I could never figure out what kind of player JT Miller was. It doesn't stand for anything.

Data used to build these files is via Corsica.Hockey, Hockey-Reference (ages of players). Similarity Score is based off gorilla math.